backend/requirements.txt
absl-py==2.3.1
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cmake==4.2.0
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mediapipe==0.10.31
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
pycparser==2.23
scikit-learn==1.8.0
scipy==1.16.3
sounddevice==0.5.3
threadpoolctl==3.6.0
Werkzeug==3.1.4
